requirements.txt
python-dotenv==1.1.1*
requests==2.32.4*
python-telegram-bot==13.0*